requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=3.6.2=pyhd8ed1ab_0
appdirs=1.4.4=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39h02fc5c5_3
asttokens=2.2.1=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
babel=2.11.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
backtesting=0.3.3=pypi_0
beautifulsoup4=4.11.1=pyha770c72_0
bleach=5.0.1=pyhd8ed1ab_0
bokeh=3.2.1=pypi_0
bottleneck=1.3.5=py39heec5a64_0
brotlipy=0.7.0=py39h02fc5c5_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2023.7.22=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2022.12.7=pypi_0
cffi=1.15.1=py39h7e6b969_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
cloudpickle=2.2.1=pypi_0
comm=0.1.2=pyhd8ed1ab_0
contourpy=1.1.0=pypi_0
cryptography=39.0.0=py39he2a39a8_0
cycler=0.11.0=pypi_0
dacite=1.8.1=pypi_0
debugpy=1.6.4=py39h23fbdae_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
flatbuffers=1.12=pypi_0
flit-core=3.8.0=pyhd8ed1ab_0
fonttools=4.40.0=pypi_0
frozendict=2.3.8=pypi_0
gast=0.4.0=pypi_0
google-auth=2.21.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.56.0=pypi_0
h5py=3.6.0=nompi_py39hd982b79_100
hdf5=1.12.1=nompi_hd9dbc9e_104
html5lib=1.1=pypi_0
htmlmin=0.1.12=pypi_0
idna=3.4=pyhd8ed1ab_0
imagehash=4.3.1=pypi_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_resources=5.10.2=pyhd8ed1ab_0
ipykernel=6.19.4=pyh736e0ef_0
ipython=8.8.0=pyhd1c38e8_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.4=pyhd8ed1ab_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=py39hca03da5_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter=1.0.0=py39h2804cbe_8
jupyter_client=7.4.8=pyhd8ed1ab_0
jupyter_console=6.4.4=pyhd8ed1ab_0
jupyter_core=5.1.2=py39h2804cbe_0
jupyter_events=0.5.0=pyhd8ed1ab_1
jupyter_server=2.0.6=pyhd8ed1ab_0
jupyter_server_terminals=0.4.3=pyhd8ed1ab_0
jupyterlab=3.5.2=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.18.0=pyhd8ed1ab_0
jupyterlab_widgets=3.0.5=pyhd8ed1ab_0
keras=2.9.0=pypi_0
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h69eda48_0
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libcurl=7.87.0=h9049daf_0
libcxx=14.0.6=h2692d47_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=11_3_0_hd922786_27
libgfortran5=11.3.0=hdaf2cc0_27
liblapack=3.9.0=16_osxarm64_openblas
libnghttp2=1.51.0=hae82a92_0
libopenblas=0.3.21=openmp_hc731615_3
libprotobuf=3.19.4=hccf11d3_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=h7a5bd25_3
libzlib=1.2.13=h03a7124_4
lightgbm=3.3.5=py39h7dbb43e_0
llvm-openmp=15.0.6=h7cfbb63_0
llvmlite=0.40.1=pypi_0
lxml=4.9.3=pypi_0
markupsafe=2.1.1=py39h02fc5c5_2
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.4=pyhd8ed1ab_0
multimethod=1.9.1=pypi_0
multitasking=0.0.11=pypi_0
nbclassic=0.4.8=pyhd8ed1ab_0
nbclient=0.7.2=pyhd8ed1ab_0
nbconvert=7.2.7=pyhd8ed1ab_0
nbconvert-core=7.2.7=pyhd8ed1ab_0
nbconvert-pandoc=7.2.7=pyhd8ed1ab_0
nbformat=5.7.1=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
notebook=6.5.2=pyha770c72_1
notebook-shim=0.2.2=pyhd8ed1ab_0
numba=0.57.1=pypi_0
numexpr=2.8.4=py39h79ee842_1
numpy=1.23.2=py39h3668e8b_0
openssl=3.1.2=h53f4e23_0
packaging=22.0=pyhd8ed1ab_0
pandas=1.5.3=py39h78102c4_0
pandoc=2.19.2=hce30654_1
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
phik=0.12.3=pypi_0
pickleshare=0.7.5=py_1003
pillow=10.0.0=pypi_0
pip=22.3.1=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=2.6.2=pyhd8ed1ab_0
pooch=1.4.0=pyhd3eb1b0_0
prettytable=3.5.0=py39hca03da5_0
prometheus_client=0.15.0=pyhd8ed1ab_0
prompt-toolkit=3.0.36=pyha770c72_0
prompt_toolkit=3.0.36=hd8ed1ab_0
protobuf=4.23.4=pypi_0
psutil=5.9.4=py39h02fc5c5_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.10.12=pypi_0
pydot=1.4.2=pypi_0
pygments=2.14.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=py39h02fc5c5_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.15=hea58f1e_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python-json-logger=2.0.4=pyhd8ed1ab_0
python_abi=3.9=3_cp39
pytz=2022.7=pyhd8ed1ab_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=py39h02fc5c5_5
pyzmq=24.0.1=py39h0553236_1
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_1
scikit-learn=1.2.2=py39h313beb8_1
scipy=1.10.1=py39h18313fe_0
seaborn=0.12.2=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=65.6.3=pyhd8ed1ab_0
shap=0.42.1=pypi_0
six=1.15.0=pypi_0
slicer=0.0.7=pypi_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=pypi_0
ta-lib=0.4.27=pypi_0
tangled-up-in-unicode=0.2.0=pypi_0
tensorboard=2.9.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorflow-deps=2.10.0=0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-macos=2.9.0=pypi_0
tensorflow-metal=0.5.0=pypi_0
terminado=0.17.1=pyhd1c38e8_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.2=py39h02fc5c5_1
tqdm=4.66.1=pypi_0
traitlets=5.8.0=pyhd8ed1ab_0
typeguard=2.13.3=pypi_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
urllib3=1.26.13=pyhd8ed1ab_0
visions=0.7.5=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.4.2=pyhd8ed1ab_0
werkzeug=2.2.2=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
widgetsnbextension=4.0.5=pyhd8ed1ab_0
wordcloud=1.9.2=pypi_0
wrapt=1.14.1=pypi_0
xgboost=1.7.6=pypi_0
xyzservices=2023.7.0=pypi_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
ydata-profiling=4.5.1=pypi_0
yfinance=0.2.23=pypi_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.13=h03a7124_4
